Rename `key_variants` module to `signatures`
Resolves: #96
